+++ .gitignore
@@ -1,0 +1,4 @@
+*.pyc
+venv/
+**/__pycache__
+config_test.json
